+++ magda-authorization-api/src/test/buildCreateApiRouter.spec.ts
@@ -13,7 +13,12 @@ import mockDatabase from "./mockDatabase";
 import mockUserDataStore from "magda-typescript-common/src/test/mockUserDataStore";
 import Database from "../Database";
 import { userDataArb } from "./arbitraries";
+import {
+    uuidArb,
+    lcAlphaNumStringArbNe
+} from "magda-typescript-common/src/test/arbitraries";
 import { Request } from "supertest";
+import mockApiKeyStore from "./mockApiKeyStore";
 
 describe("Auth api router", function (this: Mocha.ISuiteCallbackContext) {
     this.timeout(10000);
@@ -28,6 +33,7 @@ describe("Auth api router", function (this: Mocha.ISuiteCallbackContext) {
 
     afterEach(function () {
         mockUserDataStore.reset();
+        mockApiKeyStore.reset();
     });
 
     function retrieveArgv() {
@@ -356,6 +362,92 @@ describe("Auth api router", function (this: Mocha.ISuiteCallbackContext) {
                             throw e;
                         }
                     })
+                );
+            });
+        });
+    });
+
+    describe("GET /private/users/apikey/:apiKeyId", () => {
+        silenceErrorLogs(() => {
+            it("should return correct user data with correct API key & API Key Id", async () => {
+                await jsc.assert(
+                    jsc.forall(
+                        uuidArb,
+                        userDataArb,
+                        async (apiKey, userData) => {
+                            try {
+                                mockUserDataStore.reset();
+                                const {
+                                    id: userId
+                                } = mockUserDataStore.createRecord(userData);
+
+                                mockApiKeyStore.reset();
+                                const {
+                                    id: apiKeyId
+                                } = await mockApiKeyStore.create(
+                                    userId,
+                                    apiKey
+                                );
+
+                                const res = await request(app)
+                                    .get(`/private/users/apikey/${apiKeyId}`)
+                                    .set("X-Magda-API-Key", apiKey);
+
+                                expect(res.status).to.equal(200);
+                                expect(res.body).to.be.a("object");
+                                expect(res.body.id).to.be.a("string");
+                                expect({
+                                    id: userId,
+                                    ...userData
+                                }).to.deep.include(res.body);
+
+                                return true;
+                            } catch (e) {
+                                throw e;
+                            }
+                        }
+                    ),
+                    // --- we need to limit the no. of test here as bcrypt hash is very slow
+                    { tests: 5 }
+                );
+            });
+
+            it("should deny access with incorrect API key & API Key Id", async () => {
+                await jsc.assert(
+                    jsc.forall(
+                        uuidArb,
+                        lcAlphaNumStringArbNe,
+                        userDataArb,
+                        async (apiKey, randomStr, userData) => {
+                            try {
+                                mockUserDataStore.reset();
+                                const {
+                                    id: userId
+                                } = mockUserDataStore.createRecord(userData);
+
+                                mockApiKeyStore.reset();
+                                const {
+                                    id: apiKeyId
+                                } = await mockApiKeyStore.create(
+                                    userId,
+                                    apiKey
+                                );
+
+                                const res = await request(app)
+                                    .get(`/private/users/apikey/${apiKeyId}`)
+                                    // --- add non empty random string to make it incorrect key
+                                    .set("X-Magda-API-Key", apiKey + randomStr);
+
+                                expect(res.status).to.equal(401);
+
+                                return true;
+                            } catch (e) {
+                                throw e;
+                            }
+                        }
+                    ),
+                    // --- we need to limit the no. of test here as bcrypt hash is very slow
+                    { tests: 5 }
                 );
             });
         });